refact catalog api
